% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_root_url.R
\name{build_root_url}
\alias{build_root_url}
\title{Build the root URL for a .Stat Suite SDMX API}
\usage{
build_root_url(
  data_space = c("design", "collection", "processing", "staging", "disseminate",
    "disseminateext", "archive", "archiveext"),
  environment = c("qa", "prod")
)
}
\arguments{
\item{data_space}{One of the data spaces available in the .Stat instance}

\item{environment}{Either `"qa"` or `"prod"`}
}
\value{
A length‑one character vector containing the assembled URL.
}
\description{
Returns the root URL of the API service
}
\examples{
build_root_url("design", "prod")
#> "https://designnsiserviceprod.ase.worldbank.org/rest/"
}
