% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_structures.R
\name{parse_codelists}
\alias{parse_codelists}
\title{Parse all Codelists from an SDMX Structure message}
\usage{
parse_codelists(xml)
}
\arguments{
\item{xml}{`xml_document` **or** string containing the XML.}
}
\value{
A tibble (one row per `<structure:Codelist>`).
}
\description{
Parse all Codelists from an SDMX Structure message
}
