% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_structures.R
\name{parse_dataflows}
\alias{parse_dataflows}
\title{Parse all Dataflows from an SDMX Structure message}
\usage{
parse_dataflows(xml)
}
\arguments{
\item{xml}{`xml_document` **or** string containing the XML.}
}
\value{
A tibble (one row per `<structure:Dataflow>`).
}
\description{
Parse all Dataflows from an SDMX Structure message
}
