% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_structures.R
\name{parse_dsds}
\alias{parse_dsds}
\title{Parse all Data Structure Definitions (DSDs) from an SDMX Structure message}
\usage{
parse_dsds(xml)
}
\arguments{
\item{xml}{An `xml_document` **or** a single XML string.}
}
\value{
A tibble with one row per `<structure:DataStructure>` element.
}
\description{
Parse all Data Structure Definitions (DSDs) from an SDMX Structure message
}
